+++ route.go
@@ -3,31 +3,11 @@ package netlink
 import (
 	"fmt"
 	"net"
-	"strconv"
 	"strings"
-
-	"golang.org/x/sys/unix"
 )
 
 // Scope is an enum representing a route scope.
 type Scope uint8
-
-func (s Scope) String() string {
-	switch s {
-	case SCOPE_UNIVERSE:
-		return "universe"
-	case SCOPE_SITE:
-		return "site"
-	case SCOPE_LINK:
-		return "link"
-	case SCOPE_HOST:
-		return "host"
-	case SCOPE_NOWHERE:
-		return "nowhere"
-	default:
-		return "unknown"
-	}
-}
 
 type NextHopFlag int
 
@@ -49,57 +29,6 @@ type Encap interface {
 
 //Protocol describe what was the originator of the route
 type RouteProtocol int
-
-func (p RouteProtocol) String() string {
-	switch int(p) {
-	case unix.RTPROT_BABEL:
-		return "babel"
-	case unix.RTPROT_BGP:
-		return "bgp"
-	case unix.RTPROT_BIRD:
-		return "bird"
-	case unix.RTPROT_BOOT:
-		return "boot"
-	case unix.RTPROT_DHCP:
-		return "dhcp"
-	case unix.RTPROT_DNROUTED:
-		return "dnrouted"
-	case unix.RTPROT_EIGRP:
-		return "eigrp"
-	case unix.RTPROT_GATED:
-		return "gated"
-	case unix.RTPROT_ISIS:
-		return "isis"
-	//case unix.RTPROT_KEEPALIVED:
-	//	return "keepalived"
-	case unix.RTPROT_KERNEL:
-		return "kernel"
-	case unix.RTPROT_MROUTED:
-		return "mrouted"
-	case unix.RTPROT_MRT:
-		return "mrt"
-	case unix.RTPROT_NTK:
-		return "ntk"
-	case unix.RTPROT_OSPF:
-		return "ospf"
-	case unix.RTPROT_RA:
-		return "ra"
-	case unix.RTPROT_REDIRECT:
-		return "redirect"
-	case unix.RTPROT_RIP:
-		return "rip"
-	case unix.RTPROT_STATIC:
-		return "static"
-	case unix.RTPROT_UNSPEC:
-		return "unspec"
-	case unix.RTPROT_XORP:
-		return "xorp"
-	case unix.RTPROT_ZEBRA:
-		return "zebra"
-	default:
-		return strconv.Itoa(int(p))
-	}
-}
 
 // Route represents a netlink route.
 type Route struct {
